% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cd2030.popderiv-package.R
\docType{package}
\name{cd2030.popderiv-package}
\alias{cd2030.popderiv}
\alias{cd2030.popderiv-package}
\title{cd2030.popderiv: Derive Population Denominators}
\description{
Tools to derive population denominators for demographic, and coverage analyses.
}
\author{
\strong{Maintainer}: David Kariuki \email{hello@damurka.com}

}
\keyword{internal}
